simplify module utils files
