+++ awx_collection/plugins/module_utils/awxkit.py
@@ -2,7 +2,7 @@ from __future__ import absolute_import, division, print_function
 
 __metaclass__ = type
 
-from .controller_module import ControllerModule
+from .controller_api import ControllerModule
 from ansible.module_utils.basic import missing_required_lib
 
 try:
